Fixed a bug with comparison of transactionIDs:
Bug:
transactionIDs were lexicograpically compared as strings
instead of by their numerical value.

Explanation:
The usual case is that the transactionIDs of the witholding
tax and the dividend are close - meaning that they have the
same number of digits. Hence the original code worked most
of the time, because lexicograpical comparison is the same
as numerical when the numbers have the same number of digits.